shared/scripts/py/pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcc>=0.1.10",
    "qemu-qmp>=0.0.5",
]
